Move API to correct package
