--- pkg/scsstore/scs_test.go
@@ -1,9 +1,0 @@
-package scsstore
-
-import "testing"
-
-func TestThat(t *testing.T) {
-	hello()
-
-	t.Fatal("NO")
-}
